Remove generate token modal
